+++ src/yetty/cards/python/python.cpp
@@ -139,7 +139,7 @@ public:
         return Ok();
     }
 
-    void setCellSize(uint32_t cellWidth, uint32_t cellHeight) override {
+    void setCellSize(float cellWidth, float cellHeight) override {
         if (_cellWidth != cellWidth || _cellHeight != cellHeight) {
             _cellWidth = cellWidth;
             _cellHeight = cellHeight;
