+++ src/yetty/cards/ydraw/ydraw.cpp
@@ -415,7 +415,7 @@ public:
     // Events - zoom/pan + text selection
     //=========================================================================
 
-    void setCellSize(uint32_t cellWidth, uint32_t cellHeight) override {
+    void setCellSize(float cellWidth, float cellHeight) override {
         _cellWidth = cellWidth;
         _cellHeight = cellHeight;
     }
